+++ ChefMaster/Assets/Scripts/Vegetable.cs
@@ -1,10 +1,14 @@
-﻿using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
+﻿using UnityEngine;
+
+/// <summary>
+/// 
+/// This class defines the properties of a vegetable
+/// Used in vegetable prefab 
+/// 
+/// </summary>
 
 public class Vegetable : MonoBehaviour
 {
-    public new string name;
     public InventoryEnum vegetableType;
     public float choppingTime;
     public GameObject model;