+++ ChangeLog.md
@@ -1,5 +1,21 @@
 # Change Log
 
+## [0.0.2] - 19-07-2020
+### Added
+- Added Interface function.
+- UIManager class added.
+- Text mesh pro package added.
+- Basic Customer UI and Screen HUD UI added.  
+- Instruction.md file added.
+- Player_2 controls added.
+### Changed
+- README file updated.
+- Interface function is used instead of Tags for player interaction.
+- Major changes in player control script.
+-  Separate script created for individual intractable objects. 
+### Fixed
+- Item pickup and drop action bug fixed. 
+
 ## [0.0.1] - 18-07-2020
 ### Added
 - Change log file crated.